#3442 - some of the fixes
